test: move handshake tests to one file
